Virtual Screen - a virtual reality application
This system can automatically locate the four corners of an A4 paper
(with a specific pattern on each corner of the A4 paper), and actively
track these corners with a PTZ-controlled camera. The tracking error is
generally no more than 10 pixels. With this tracking mechanism, the
system presents us a virtual screen using the A4 paper. Even in
relatively poor light conditions, it can achieve accurate recognition
and tracking of the patterns on the paper.